Add support for sending command as update
